reorganized account settings
