+++ auth-web/src/components/auth/account-settings/account-info/AccountInfo.vue
@@ -143,7 +143,7 @@ import { AccountSettings } from '@/models/account-settings'
 import { Address } from '@/models/address'
 import BaseAddressForm from '@/components/auth/common/BaseAddressForm.vue'
 import ConfigHelper from '@/util/config-helper'
-import OrgAdminContact from '@/components/auth/OrgAdminContact.vue'
+import OrgAdminContact from '@/components/auth/account-settings/account-info/OrgAdminContact.vue'
 import OrgModule from '@/store/modules/org'
 import { PaymentSettings } from '@/models/PaymentSettings'
 import { addressSchema } from '@/schemas'
