+++ auth-web/src/components/auth/account-settings/team-management/AnonymousUserManagement.vue
@@ -232,11 +232,11 @@ import {
 } from '@/models/Organization'
 import { Component, Mixins, Prop } from 'vue-property-decorator'
 import { mapActions, mapState } from 'vuex'
-import AddUsersForm from '@/components/auth/AddUsersForm.vue'
-import AddUsersSuccess from '@/components/auth/AddUsersSuccess.vue'
+import AddUsersForm from '@/components/auth/account-settings/team-management/AddUsersForm.vue'
+import AddUsersSuccess from '@/components/auth/account-settings/team-management/AddUsersSuccess.vue'
 import MemberDataTable from '@/components/auth/account-settings/team-management/MemberDataTable.vue'
 import ModalDialog from '@/components/auth/common/ModalDialog.vue'
-import PasswordReset from '@/components/auth/PasswordReset.vue'
+import PasswordReset from '@/components/auth/account-settings/team-management/PasswordReset.vue'
 import { SearchFilterCodes } from '@/util/constants'
 import SearchFilterInput from '@/components/auth/common/SearchFilterInput.vue'
 import { SearchFilterParam } from '@/models/searchfilter'
