+++ auth-web/src/components/auth/account-settings/team-management/UserManagement.vue
@@ -174,7 +174,7 @@ import AccountChangeMixin from '@/components/auth/mixins/AccountChangeMixin.vue'
 import AccountMixin from '@/components/auth/mixins/AccountMixin.vue'
 import { Invitation } from '@/models/Invitation'
 import InvitationsDataTable from '@/components/auth/account-settings/team-management/InvitationsDataTable.vue'
-import InviteUsersForm from '@/components/auth/InviteUsersForm.vue'
+import InviteUsersForm from '@/components/auth/account-settings/team-management/InviteUsersForm.vue'
 import { KCUserProfile } from 'sbc-common-components/src/models/KCUserProfile'
 import MemberDataTable from '@/components/auth/account-settings/team-management/MemberDataTable.vue'
 import ModalDialog from '@/components/auth/common/ModalDialog.vue'
