+++ auth-web/src/components/auth/create-account/AccountCreateBasic.vue
@@ -72,7 +72,6 @@ import { Member, Organization } from '@/models/Organization'
 import { mapActions, mapMutations, mapState } from 'vuex'
 import { Address } from '@/models/address'
 import BaseAddressForm from '@/components/auth/common/BaseAddressForm.vue'
-import BcolLogin from '@/components/auth/BcolLogin.vue'
 import ConfirmCancelButton from '@/components/auth/common/ConfirmCancelButton.vue'
 import OrgModule from '@/store/modules/org'
 import Steppable from '@/components/auth/common/stepper/Steppable.vue'
@@ -81,7 +80,6 @@ import { getModule } from 'vuex-module-decorators'
 
 @Component({
   components: {
-    BcolLogin,
     BaseAddressForm,
     ConfirmCancelButton
   },
