+++ auth-web/src/components/auth/mixins/TeamManagementMixin.vue
@@ -11,7 +11,6 @@ import { Event } from '@/models/event'
 import { EventBus } from '@/event-bus'
 import { Invitation } from '@/models/Invitation'
 import InvitationsDataTable from '@/components/auth/account-settings/team-management/InvitationsDataTable.vue'
-import InviteUsersForm from '@/components/auth/InviteUsersForm.vue'
 import { KCUserProfile } from 'sbc-common-components/src/models/KCUserProfile'
 import ModalDialog from '@/components/auth/common/ModalDialog.vue'
 import OrgModule from '@/store/modules/org'
